requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.13.3
python-multipart==0.0.9
httpx==0.27.2
alembic==1.16.5
openai==2.32.0
pytest==8.3.5
SQLAlchemy==2.0.49
python-dotenv==1.0.1
langgraph==1.1.9
langchain-core==1.3.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
APScheduler==3.10.4
oss2==2.19.1
